Lee Elliott:
Here're some updates for some of the instruments, the TSR2 and the B52.

The instruments are just parameterised/non-specific versions of the engine
and fuel gauges.  The old gauges will still be used by the a10s and the sea
hawks until I get them done.

There're a couple of tweaks to the tsr2 yasim config and amendments to the
model file to change the angle at which the airbrakes operate, correct the
direction that the nose gear retracts/extends and to rotate the main-gear
carts during retraction/extention.  There're also new panels, using the new
instruments.

The b52 yasim config has some big changes due to finding some more info i.e.
wing incidence of 6 degrees, fuel capacities and aileron changes (removed
from G & H versions!).  It's now based on an 'F' model and I've put some
comments in the yasim config about it.  The b52-readme.txt includes a
suggested method on getting airborne;).   I've included an amended model -
B52-F.3ds, which is the default in the new model file, and a couple of panels.

I'm quite pleased with the b52 - the take-offs resemble the photographs I've
seen and I think I've got reasonable values where I've had to guess at stuff,
but the results from the yasim solver look as through the model is
approaching the limits of acceptability for yasim.  By this I mean that the
model is probably a bit dodgy, not that the yasim solver is limited.