Changes from Dave Perry:
Here is a patch that included the changes in my 2/22/06 posting plus the
3d-instrument light "emission" animations.  I also added a magnetic
compas mount and moved it forward and down to better match the real ac
and make it easier to read when adjusting the hi.  Ignore the "?" flaged
files at the beginning.  none are required.  I noticed that there is a
backup copy of pa24-250.ac in the pa24-250 folder in cvs.  It can be
deleted.  The only copy of that file required is in pa24-25/Models.

I "flew" a shake down instrument XC from 2v2 (Longmont, CO) to KLVN
(Airlake, just south of Minneapolis, MN).  Even the fuel management is
realistic.  The autopilot on the menu bar works well.  Give the changes
a try and enjoy!